requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_llvm
blas=1.0=mkl
brotlipy=0.7.0=py37h5e8e339_1001
c-ares=1.17.1=h7f98852_1
ca-certificates=2021.5.30=ha878542_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
certifi=2021.5.30=py37h89c1867_0
cffi=1.14.5=py37hc58025e_0
chardet=4.0.0=py37h89c1867_1
cryptography=3.4.7=py37h5d9358c_0
cudatoolkit=11.0.3=h15472ef_8
h5py=3.1.0=nompi_py37h1e651dc_100
hdf5=1.10.6=nompi_h6a2412b_1114
idna=2.10=pyh9f0ad1d_0
krb5=1.17.2=h926e7f8_0
ld_impl_linux-64=2.35.1=hea4e1c9_2
libcurl=7.75.0=hc4aaa36_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libffi=3.3=h58526e2_2
libgcc-ng=9.3.0=h2828fa1_18
libgfortran-ng=9.3.0=hff62375_18
libgfortran5=9.3.0=hff62375_18
libnghttp2=1.43.0=h812cca2_0
libssh2=1.9.0=ha56f1ee_6
libstdcxx-ng=9.3.0=h6de172a_18
libuv=1.41.0=h7f98852_0
llvm-openmp=11.1.0=h4bd325d_0
mkl=2020.4=h726a3e6_304
mkl-service=2.3.0=py37h8f50634_2
mkl_fft=1.3.0=py37h902c9e0_1
mkl_random=1.2.0=py37h9fdb41a_1
ncurses=6.2=h58526e2_4
ninja=1.10.2=h4bd325d_0
numpy=1.19.2=py37h54aff64_0
numpy-base=1.19.2=py37hfa32c7d_0
openssl=1.1.1k=h7f98852_0
pip=21.0.1=pyhd8ed1ab_0
pycparser=2.20=pyh9f0ad1d_2
pyopenssl=20.0.1=pyhd8ed1ab_0
pysocks=1.7.1=py37h89c1867_3
python=3.7.9=h7579374_0
python_abi=3.7=1_cp37m
pytorch=1.7.1=py3.7_cuda11.0.221_cudnn8.0.5_0
readline=8.0=he28a2e2_2
requests=2.25.1=pyhd3deb0d_0
setuptools=49.6.0=py37h89c1867_3
six=1.15.0=pyh9f0ad1d_0
sqlite=3.35.2=h74cdb3f_0
tk=8.6.10=h21135ba_1
torchtext=0.8.1=py37
tqdm=4.59.0=pyhd8ed1ab_0
typing_extensions=3.7.4.3=py_0
urllib3=1.26.5=pyhd8ed1ab_0
wheel=0.36.2=pyhd3deb0d_0
xz=5.2.5=h516909a_1
zlib=1.2.11=h516909a_1010
